chore: restructure verifier as part of sdk
